pyproject.toml
[tool.poetry]
name = "line-provider"
version = "0.1.0"
description = ""
authors = ["videbul <videbul@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.114.1"
pydentic = "^0.0.1.dev3"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.34"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

